+++ rbac-distributed-service-impl/src/main/java/com/dwarfeng/rbacds/impl/dao/preset/PexpPresetCriteriaMaker.java
@@ -4,6 +4,7 @@ import com.dwarfeng.rbacds.stack.service.PexpMaintainService;
 import com.dwarfeng.subgrade.sdk.hibernate.criteria.PresetCriteriaMaker;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import org.hibernate.criterion.DetachedCriteria;
+import org.hibernate.criterion.MatchMode;
 import org.hibernate.criterion.Restrictions;
 import org.springframework.stereotype.Component;
 
@@ -23,6 +24,12 @@ public class PexpPresetCriteriaMaker implements PresetCriteriaMaker {
                 break;
             case PexpMaintainService.PEXP_FOR_ROLE_SET:
                 childForRoleSet(detachedCriteria, objects);
+                break;
+            case PexpMaintainService.PEXP_FOR_ROLE_DESCRIPTION_ASC:
+                pexpForRoleDescriptionAsc(detachedCriteria, objects);
+                break;
+            case PexpMaintainService.PEXP_FOR_ROLE_DESCRIPTION_LIKE_DESCRIPTION_ASC:
+                pexpForRoleDescriptionLikeDescriptionAsc(detachedCriteria, objects);
                 break;
             default:
                 throw new IllegalArgumentException("无法识别的预设: " + s);
@@ -55,6 +62,36 @@ public class PexpPresetCriteriaMaker implements PresetCriteriaMaker {
                     detachedCriteria.add(Restrictions.in("roleId", stringList(stringIdKeys)));
                 }
             }
+        } catch (Exception e) {
+            throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
+        }
+    }
+
+    private void pexpForRoleDescriptionAsc(DetachedCriteria detachedCriteria, Object[] objects) {
+        try {
+            if (Objects.isNull(objects[0])) {
+                detachedCriteria.add(Restrictions.isNull("roleId"));
+            } else {
+                StringIdKey stringIdKey = (StringIdKey) objects[0];
+                detachedCriteria.add(Restrictions.eqOrIsNull("roleId", stringIdKey.getStringId()));
+            }
+            detachedCriteria.addOrder(org.hibernate.criterion.Order.asc("description"));
+        } catch (Exception e) {
+            throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
+        }
+    }
+
+    private void pexpForRoleDescriptionLikeDescriptionAsc(DetachedCriteria detachedCriteria, Object[] objects) {
+        try {
+            if (Objects.isNull(objects[0])) {
+                detachedCriteria.add(Restrictions.isNull("roleId"));
+            } else {
+                StringIdKey stringIdKey = (StringIdKey) objects[0];
+                detachedCriteria.add(Restrictions.eqOrIsNull("roleId", stringIdKey.getStringId()));
+            }
+            String pattern = (String) objects[1];
+            detachedCriteria.add(Restrictions.like("description", pattern, MatchMode.ANYWHERE));
+            detachedCriteria.addOrder(org.hibernate.criterion.Order.asc("description"));
         } catch (Exception e) {
             throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
         }